requirements.txt
pyautogen==0.2.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
